+++ chatservermq.js
@@ -12,7 +12,17 @@ var observer = require('./observer');
 var startServer = observer.startServer;
 
 
-let active_queues_env_string = process.env.ACTIVE_QUEUES;
+let active_queues_env_string = null;
+
+if (process.env.ACTIVE_QUEUES) {
+  logger.debug("Found process.env.ACTIVE_QUEUES:", process.env.ACTIVE_QUEUES);
+  if (process.env.ACTIVE_QUEUES.trim().toLowerCase() === "none") {
+    active_queues_env_string = "none";
+  }
+  else {
+    active_queues_env_string = process.env.ACTIVE_QUEUES;
+  }
+}
 // let ALL_SUBSCRIPTIONS_TOPICS = [
 //   'outgoing',
 //   'update',
@@ -38,29 +48,35 @@ let ALL_SUBSCRIPTIONS_TOPICS = {
 };
 // let ALL_SUBSCRIPTIONS_TOPICS_AS_KEYS;
 // ALL_SUBSCRIPTIONS_TOPICS.forEach(e => {ALL_SUBSCRIPTIONS_TOPICS_AS_KEYS[e] = true});
-if (!active_queues_env_string) {
+if (active_queues_env_string === null) {
   active_queues = ALL_SUBSCRIPTIONS_TOPICS;
-  logger.log("active_queues unset. Using default:", active_queues);
+  logger.info("(Observer) ACTIVE_QUEUES unset. Using default (all: messages & persist):", active_queues);
+}
+else if (active_queues_env_string === "none") {
+  logger.error("(Observer) Not ACTIVE_QUEUES enabled. Going on anyway...");
+  logger.error("(Observer) To activate queues use syntax: active_queues=messages,persist (valid topics: messages | persist | messages,persist)");
 }
 else {
   active_queues_env_array = active_queues_env_string.split(",");
-  logger.log("active_queues found:", active_queues_env_array)
+  logger.info("(Observer) ACTIVE_QUEUES found:", active_queues_env_array)
   if (active_queues_env_array.length == 0) {
-    logger.error("active_queues length cant be ZERO.");
-    logger.error("Use syntax: active_queues=outgoing,persist (valid topics: outgoing,update,persist,archive,presence,update_group,delivered)");
+    logger.error("(Observer) Not ACTIVE_QUEUES enabled.");
+    logger.error("(Observer) To activate queues use syntax: active_queues=messages,persist (valid topics: messages | persist | messages,persist)");
     process.exit(1);
   }
-  active_queues_env_array.forEach((e,i) => {
-    topic = e.trim();
-    if (!ALL_SUBSCRIPTIONS_TOPICS[topic]) {
-      logger.error("Invalid subscription topic:", topic);
-      logger.error("Use syntax: active_queues=outgoing,persist (valid topics: outgoing,update,persist,archive,presence,update_group,delivered)");
-      process.exit(1);
-    }
-    else {
-      active_queues[topic] = true;
-    }
-  });
+  else {
+    active_queues_env_array.forEach((e,i) => {
+      topic = e.trim();
+      if (!ALL_SUBSCRIPTIONS_TOPICS[topic]) {
+        logger.error("Invalid subscription topic:", topic);
+        logger.error("Use syntax: active_queues=outgoing,persist (valid topics: outgoing,update,persist,archive,presence,update_group,delivered)");
+        process.exit(1);
+      }
+      else {
+        active_queues[topic] = true;
+      }
+    });
+  }
 }
 
 var webhook_enabled = process.env.WEBHOOK_ENABLED;
@@ -84,20 +100,32 @@ if (process.env.WEBHOOK_EVENTS) {
   const webhook_events = process.env.WEBHOOK_EVENTS;
   webhook_events_array = webhook_events.split(",");
 }
-logger.info("webhook_events_array: " , webhook_events_array);
+logger.info("(Observer) webhook_events_array: " , webhook_events_array);
 
 let presence_enabled = false;
+logger.debug("process.env.PRESENCE_ENABLED:", process.env.PRESENCE_ENABLED)
 if (process.env.PRESENCE_ENABLED && process.env.PRESENCE_ENABLED === "true") {
   presence_enabled = true;
-  logger.info("Presence manager enabled.");
+  logger.debug("(Observer) Presence manager enabled.");
 }
 else {
-  logger.info("Presence manager disabled.");
+  logger.debug("(Observer) Presence manager disabled.");
+}
+
+let durable_enabled = true;
+logger.debug("process.env.DURABLE_ENABLED:", process.env.DURABLE_ENABLED)
+if (process.env.DURABLE_ENABLED && process.env.DURABLE_ENABLED === "true") {
+  durable_enabled = true;
+  logger.info("(Observer) Durable enabled.");
+}
+else {
+  logger.info("(Observer) Durable disabled.");
 }
 
 logger.info("Starting observer")
 async function start() {
   observer.setPresenceEnabled(presence_enabled);
+  observer.setDurableEnabled(durable_enabled);
   observer.setWebHookEnabled(webhook_enabled);
   observer.setWebHookEndpoints(webhook_endpoints_array);
   observer.setWebHookEvents(webhook_events_array);
@@ -114,7 +142,16 @@ async function start() {
   }
   // observer.setPersistentMessages(true);
 
-  await startServer({app_id: process.env.APP_ID, exchange: 'amq.topic', rabbitmq_uri:process.env.RABBITMQ_URI, mongo_uri: process.env.MONGODB_URI});
+  await startServer({
+    app_id: process.env.APP_ID,
+    exchange: 'amq.topic',
+    rabbitmq_uri:process.env.RABBITMQ_URI,
+    mongo_uri: process.env.MONGODB_URI,
+    redis_enabled: process.env.REDIS_ENABLED,
+    redis_host: process.env.REDIS_HOST,
+    redis_port: process.env.REDIS_PORT,
+    redis_password: process.env.REDIS_PASSWORD
+  });
 }
 start();