+++ mqttclient/chat21client.js
@@ -915,6 +915,7 @@ class Chat21Client {
             }
         }
         else {
+            console.log("Axios call...")
             axios(
                 {
                   url: options.url,
@@ -924,6 +925,7 @@ class Chat21Client {
                 })
               .then(function (response) {
                 if (log) {console.log("response.status:", response.status);}
+                if (log) {console.log("response.data:", response.data);}
                 if (callback) {
                     callback(null, response.headers, response.data);
                 }
@@ -992,7 +994,7 @@ class Chat21Client {
         }
         if (this.log) {console.log("starting mqtt connection with LWT on:", this.presence_topic, this.endpoint)}
         // client = mqtt.connect('mqtt://127.0.0.1:15675/ws',options)
-        console.log("starting mqtt connection with LWT on:", this.presence_topic, this.endpoint)
+        //console.log("starting mqtt connection with LWT on:", this.presence_topic, this.endpoint)
         this.client = mqtt.connect(this.endpoint,options)
         
         this.client.on('connect', // TODO if token is wrong it must reply with an error!