+++ start_messaging.sh
@@ -1,1 +1,1 @@
-ACTIVE_QUEUES=messages node chatservermq.js MSG
+ACTIVE_QUEUES=messages node chatservermq.js